+++ meshroom/nodes/general/GetMgSceneParams.py
@@ -2,8 +2,6 @@
 
 __version__ = "1.0"
 
-import json
-import logging
 import shlex
 from pathlib import Path
 from meshroom.core import desc
@@ -104,7 +102,8 @@ class GetMeshroomSceneParams(desc.CommandLineNode):
                     value=True,
                     invalidate=False
                 ),
-            ]
+            ],
+            advanced=True,
         )
     ]
 
@@ -116,94 +115,3 @@ class GetMeshroomSceneParams(desc.CommandLineNode):
             value="{nodeCacheFolder}/values.json",
         )
     ]
-
-
-class UnwrapMeshroomSceneParam(desc.InputNode, desc.InitNode):
-    """Unwrap the JSON file created by a MeshroomSceneParameter node
-    to expose its items to the graph.
-    
-    It uses the jsonFile connection to fetch parameters from the
-    MeshroomSceneParameter node to fill the selectedParameter choices,
-    so it's better to use it directly after this node.
-    """
-
-    category = "Utils"
-
-    inputs = [
-        desc.File(
-            name="jsonFile",
-            label="JSON file",
-            description=(
-                "JSON file generated by a MeshroomSceneParameter node.\n"
-                "Note that the input must be a MeshroomSceneParameter node "
-                "because we use the input to fetch the choice parameters."
-            ),
-            value="",
-            exposed=True,
-        ),
-        desc.ChoiceParam(
-            name="selectedParameter",
-            description="Parameter to fetch from the JSON file.",
-            value="",
-            values=[],
-            exclusive=True,
-            exposed=True,
-        )
-    ]
-
-    outputs = [
-        desc.StringParam(
-            name="outputValue",
-            description="Extracted value from the JSON file",
-            value="",
-        ),
-    ]
-
-    @staticmethod
-    def updateChoices(node):
-        if not node.jsonFile.isLink:
-            node.selectedParameter.setValues([])
-            return
-        inputNode = node.jsonFile.inputLink.node
-        if inputNode.nodeType != "GetMeshroomSceneParams":
-            node.selectedParameter.setValues([])
-            return
-        inputChoices = []
-        for item in inputNode.parameters.value:
-            nodeName = item.nodeInstance.value.strip()
-            paramPath = item.paramName.value.strip()
-            if nodeName and paramPath:
-                inputChoices.append(f"{nodeName}:{paramPath}")
-
-        node.selectedParameter.setValues(inputChoices)
-
-    @staticmethod
-    def setOutput(node):
-        if not node.selectedParameter or not node.selectedParameter.value:
-            return
-        selectedParameter = node.selectedParameter.value
-        logging.debug(f"[UnwrapMeshroomSceneParam] Selected parameter: {selectedParameter}")
-        nodeInstance, param = selectedParameter.split(":", 1)
-        jsonFile = node.jsonFile.value
-        if not Path(jsonFile).exists():
-            node.outputValue.value = ""
-            return
-        with open(jsonFile, "r") as f:
-            data = json.load(f)
-        logging.debug(f"JSON file : {jsonFile}:\n{data}")
-        for item in data:
-            if item.get("node") == nodeInstance and item.get("parameter") == param:
-                node.outputValue.value = item.get("value")
-                break
-        else:
-            node.outputValue.value = ""
-
-    def update(self, node):
-        try:
-            self.updateChoices(node)
-        except Exception as e:
-            logging.warning(f"[UnwrapMeshroomSceneParam] Failed to set the choices: {e}")
-        try:
-            self.setOutput(node)
-        except Exception as e:
-            logging.warning(f"[UnwrapMeshroomSceneParam] Failed to set the node output: {e}")
